+++ proyectoPokemon/src/main/java/com/adrianLopez/proyectoPokemon/dto/PokemonDTO.java
@@ -13,8 +13,8 @@ public class PokemonDTO {
 
     private int id;
     private String name;
-    private double height;
-    private double weight;
+    private int height;
+    private int weight;
     private int exp;
 
     private List<SlotPokemonDTO> slotPokemonDTOs;
