+++ proyectoPokemon/src/main/java/com/adrianLopez/proyectoPokemon/peristence/model/PokemonEntity.java
@@ -19,8 +19,8 @@ public class PokemonEntity {
 
     private int id;
     private String name;
-    private double height;
-    private double weight;
+    private int height;
+    private int weight;
     private int exp;
 
     private List<SlotPokemonEntity> slotPokemonEntities;
